// src/environments/interfaces/IEnvironment.interfaces.ts
export interface IEnvironment {
  production: boolean;
  env: string;
  CAS_SERVER_URL: string;
  REDIRECT_URI: string;
  SERVICIO_WEB: string;
  VALIDATE: string;
  URL_MICROSOFT: string;
  FOTO_TTHH: string;
  FOTO_ACADEMICO: string;
  URL_CENTRAL: string;
  CARGO_DEPENCENCIA: string;
  RUTA_TOKEN_TTHH: string;
}